組件切割
