+++ src/Nav.vue
@@ -1,0 +1,10 @@
+<script setup>
+</script>
+
+<template>
+    <nav>基本的導覽列</nav>
+</template>
+
+<style scoped>
+    nav { background: skyblue; }
+</style>
